Fixed 1 bin mismatches in ExtractSimLCs
Sometimes rounding causes one of the new, simultaneous light curves to be one bin longer than the other. If this is the case, and it's only one bin longer, work out whether the first or last bin times match more closely and remove the extra bin from the other end of the longer light curve.