ENTResponse updates
Implemented rebinning in time and energy, extracting sections of the response in time or energy, plotting image of response function, producing lag spectra (frequency and energy), converting to an EnergyLCList object .

Added new routines to LogBinning to lookup which bin a value falls into.

Added LinearBinning class for use with ENTResponse.